requirements.txt
streamlit==1.29.0
mysql-connector-python==8.2.0
